+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 4.0.1 (Unreleased)
+* Fix anonymous classes not being subject to various linting inspections
+* [Fix error "Call to a member function toString() on null" inside anonymous classes](https://gitlab.com/Serenata/Serenata/issues/203)
 * [Autocompletion doesn't add imports when function parameter name is missing](https://gitlab.com/Serenata/Serenata/issues/204)
+
 
 ## 4.0.0
 * PHP Integrator is now called Serenata